fix leds import and list slicing type

+++ src/modules/leds.py
@@ -1,4 +1,5 @@
-from opc import Client
+from .opc import Client
+# from opc import Client
 import logging
 import numpy as np 
 from utils.constants import LED_HUE_THRESHOLD_LOWER, LED_HUE_THRESHOLD_UPPER
@@ -48,7 +49,7 @@ class Leds:
             new_profile_length = profile_size - profile_remain
             profile = profile[0:new_profile_length]
         
-        trip = len(profile) / 3
+        trip = int(len(profile) / 3)
         
         red_val = profile[0:trip+1]
         green_val = profile[trip+1:trip*2+1]
